moved api request code to src folder
