+++ demo/package.json
@@ -18,7 +18,6 @@
     "@types/react-dom": "^17.0.9",
     "@types/react-hot-loader": "^4.1.1",
     "@types/react-router-dom": "^5.3.3",
-    "axios": "^1.4.0",
     "framer-motion": "^4.1.17",
     "moment": "^2.29.1",
     "react-circular-progressbar": "^2.1.0",
